refactor: move root threading functions to their own file
