+++ Makefile.am
@@ -82,6 +82,7 @@ watchman_SOURCES = \
 	root/resolve.c       \
 	root/stat.c       \
 	root/symlink.c       \
+	root/threading.c       \
 	root/vcs.c       \
 	root/watchlist.c       \
 	state.c      \
