+++ root.c
@@ -425,57 +425,6 @@ void w_root_addref(w_root_t *root)
   w_refcnt_add(&root->refcnt);
 }
 
-static bool start_detached_root_thread(w_root_t *root, char **errmsg,
-    void*(*func)(void*), pthread_t *thr) {
-  pthread_attr_t attr;
-  int err;
-
-  pthread_attr_init(&attr);
-  pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_DETACHED);
-
-  w_root_addref(root);
-  err = pthread_create(thr, &attr, func, root);
-  pthread_attr_destroy(&attr);
-
-  if (err == 0) {
-    return true;
-  }
-
-  ignore_result(asprintf(errmsg,
-        "failed to pthread_create: %s\n", strerror(err)));
-  w_root_delref(root);
-  return false;
-}
-
-bool root_start(w_root_t *root, char **errmsg)
-{
-  if (!start_detached_root_thread(root, errmsg,
-        run_notify_thread, &root->notify_thread)) {
-    return false;
-  }
-
-  // Wait for it to signal that the watcher has been initialized
-  w_pending_coll_lock_and_wait(&root->pending, -1 /* infinite */);
-  w_pending_coll_unlock(&root->pending);
-
-  if (!start_detached_root_thread(root, errmsg,
-        run_io_thread, &root->io_thread)) {
-    w_root_cancel(root);
-    return false;
-  }
-  return true;
-}
-
-void signal_root_threads(w_root_t *root) {
-  // Send SIGUSR1 to interrupt blocking syscalls on the
-  // worker threads.  They'll self-terminate.
-  if (!pthread_equal(root->notify_thread, pthread_self())) {
-    pthread_kill(root->notify_thread, SIGUSR1);
-  }
-  w_pending_coll_ping(&root->pending);
-  root->watcher_ops->root_signal_threads(root);
-}
-
 void w_root_schedule_recrawl(w_root_t *root, const char *why)
 {
   if (!root->should_recrawl) {
